# Speed up adler32.Write