Write() used to allocate both a full copy of the passed data and a new vanilla hasher. This removes both allocations in the common case and increases throughput by about 36% as a result. It makes this adler32.Write() comparable in performance to the vanilla hash/adler32.Write().

            benchmark              old ns/op     new ns/op     delta
            BenchmarkWriteKB-8     647           476           -26.43%
    
            benchmark              old MB/s     new MB/s     speedup
            BenchmarkWriteKB-8     1582.46      2148.24      1.36x
    
            benchmark              old allocs     new allocs     delta
            BenchmarkWriteKB-8     2              0              -100.00%
    
            benchmark              old bytes     new bytes     delta
            BenchmarkWriteKB-8     1028          0             -100.00%