@chmduquesne 2017-01-18
Thanks for the awesome PR! Merged.