@chmduquesne 2017-01-18
I did not realize that this lib was a dependency of syncthing!

You might want to pull it as

    gopkg.in/chmduquesne/rollinghash.v2/adler32

instead of

    github.com/chmduquesne/rollinghash/adler32

This repo respects semantic versioning, so you'll be protected against unexpected changes in the interface.