@calmh 2017-01-18
Thanks. I don't especially like gopkg.in and we use vendoring so we'll stick to the original import path as long as we are allowed to. :)