@chmduquesne 2017-01-18
No worries, feel free to handle your dependencies how you like :-). As much as I find gopkg.in convenient, I certainly don't want to force anyone to use it.